fix: add browser util for fix review issue
