+++ packages/web3/src/connect-button/__tests__/index.test.tsx
@@ -2,6 +2,7 @@ import { ConnectButton } from '..';
 import { fireEvent, render } from '@testing-library/react';
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { mockClipboard } from '../../utils/test-utils';
+import { readCopyText } from '../../utils';
 
 describe('ConnectButton', () => {
   let resetMockClipboard: () => void;
@@ -108,9 +109,7 @@ describe('ConnectButton', () => {
       expect(baseElement.querySelector('.ant-message-notice-content')?.textContent).toBe(
         'Address Copied!',
       );
-      expect(navigator.clipboard.readText()).resolves.toBe(
-        '0x3ea2cfd153b8d8505097b81c87c11f5d05097c18',
-      );
+      expect(readCopyText()).resolves.toBe('0x3ea2cfd153b8d8505097b81c87c11f5d05097c18');
     });
   });
 
@@ -129,7 +128,7 @@ describe('ConnectButton', () => {
       expect(baseElement.querySelector('.ant-message-notice-content')?.textContent).toBe(
         'Address Copied!',
       );
-      expect(navigator.clipboard.readText()).resolves.toBe('aaaaaabbbbbbcccccc');
+      expect(readCopyText()).resolves.toBe('aaaaaabbbbbbcccccc');
     });
   });
 });
